requirements.txt
langchain 
langchain-community 
langchain-openai 
pinecone 
pinecone-client 
flask 
twilio 
python-dotenv